+++ tests/functional/chalk/runner.py
@@ -357,6 +357,7 @@ class Chalk:
         stdin: Optional[bytes] = None,
         tty: bool = False,
         show_config: bool = False,
+        inject_binary_into_zip: bool = False,
     ) -> ChalkProgram:
         params = params or []
         cmd: list[str] = [str(self.binary)]
@@ -388,6 +389,8 @@ class Chalk:
             cmd += ["--debug"]
         if no_color:
             cmd += ["--no-color"]
+        if inject_binary_into_zip:
+            cmd += ["--inject-binary-into-zip"]
         if params:
             cmd += params
 
@@ -446,6 +449,7 @@ class Chalk:
         expecting_report: bool = True,
         expecting_chalkmarks: bool = True,
         use_embedded: bool = True,
+        inject_binary_into_zip=False,
     ) -> ChalkProgram:
         result = self.run(
             command="insert",
@@ -457,6 +461,7 @@ class Chalk:
             ignore_errors=ignore_errors,
             expecting_report=expecting_report,
             use_embedded=use_embedded,
+            inject_binary_into_zip=inject_binary_into_zip,
         )
         if expecting_report:
             if expecting_chalkmarks: